% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Classify an image using an Inception v3 model}
\usage{
classify(images, model, labels, verbose = TRUE)
}
\arguments{
\item{images}{Character vector of paths, or a single path, to the image/s to be classified}

\item{model}{Character giving the path to the model file. This file is created by
\code{retrainInception}, which returns the path.}

\item{labels}{Character giving the path to the labels file. This file is created by
\code{retrainInception}, which returns the path.}

\item{verbose}{Logical should updates on classifiation be printed to console}
}
\value{
A list of dataframes, in each case giving the top 5 classifications and their
probabilities
}
\description{
Classify uses a model and labels created through training of the inception model
to assign classification probabilities to a new image or images.
}
\details{
This code requires tensorflow and python to run. These should be installed first. 
Tensorflow can be installed from within R using \code{tensorflow::install_tensorflow()}.
The function \code{classify} is a simple wrapper around pre-existing python scripts which can be found
here \url{https://github.com/tensorflow/tensorflow/tree/master/tensorflow/examples/image_retraining}
}
\examples{
\dontrun{
# Get example images (218MB)
tDir <- tempdir()
download.file(url = "http://download.tensorflow.org/example_images/flower_photos.tgz", 
              destfile = file.path(tDir, 'images.tgz'))

# Unzip example images
untar(tarfile = file.path(tDir, 'images.tgz'), exdir = file.path(tDir, 'images'))

# Train a model using images shipped with Rinception
# This is quick and dirty
incep <- retrainInception(imageDir = file.path(tDir, 'images/flower_photos'),
                          trainingSteps = 50,
                          trainingBatchSize = 10,
                          testBatchSize = 10,
                          validationBatchSize = 20)

# Get the paths to three images
images <- list.files(file.path(tDir, 'images/flower_photos'),
                     recursive = TRUE, full.names = TRUE)

# Classify 5 images at random
classify(images = sample(images, 5),
         model = incep$model,
         labels = incep$labels)

}
}
