% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rinception.R
\docType{package}
\name{Rinception}
\alias{Rinception}
\alias{Rinception-package}
\title{Rinception}
\description{
A package for creating image classifiers using transfer
learning with inception v3 architecture image recognition model.
This package is a simple wrapper around pre-existing python scripts
which can be found here \url{https://github.com/tensorflow/tensorflow/tree/master/tensorflow/examples/image_retraining}.
}
